updated license, allow for generic input to filter
