+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Changes are grouped as follows
 ### Added
 
 - Added possibility to directly insert pd.DataFrame, pd.Series and dictionaries into Clarify. NB! Infers timestamp column and only supports when there only exists one timestamp column.
+- Added possibility to directly insert dictionaries to `pyclarify.client.data_frame`, `pyclarify.client.select_signals`, and `pyclarify.client.select_items`.
 
 ### Fixed
