+++ README.md
@@ -6,16 +6,29 @@
 
 [![PyPI package](https://img.shields.io/badge/pip%20install-pyclarify-brightgreen)](https://pypi.org/project/pyclarify/)
 [![Version number](https://img.shields.io/pypi/v/pyclarify?color=green&label=version)](https://pypi.org/project/pyclarify/)
-[![PyPI - Downloads](https://img.shields.io/pypi/dm/pyclarify)](https://pypi.org/project/pyclarify/)
+[![Downloads](https://static.pepy.tech/badge/pyclarify)](https://pepy.tech/project/pyclarify)
+[![Downloads](https://static.pepy.tech/personalized-badge/pyclarify?period=month&units=none&left_color=grey&right_color=blue&left_text=downloads/month)](https://pepy.tech/project/pyclarify)
 [![Actions Status](https://github.com/clarify/pyclarify/workflows/Build%20status/badge.svg)](https://github.com/clarify/pyclarify/actions)
 [![License](https://img.shields.io/github/license/clarify/pyclarify)](https://github.com/clarify/pyclarify/blob/main/LICENSE)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/ambv/black)
 
 <hr/>
 
+```python
+from pyclarify import Client
+
+data = {
+  "times": ["2022-10-10T00:00:00"],
+  "temperature": [19]
+}
+
+client = Client("credentials.json")
+client.insert(data)
+```
+
 PyClarify helps users of Clarify to easily read, write and manipulate data in Clarify.
 
-- Data scientists can easily filter data, convert it to pandas with our built in methods, and write results back.
+- Data scientists can easily filter data, convert it to `pandas` with our built in methods, and write results back.
 - System integrators can set up pipelines for automatic streaming of data, and update labels on the fly.
 
 # Useful tutorials and documentation
